generative-ai/coursera-references/langfuse-mcp-python-main/pyproject.toml
[project]
name = "langfuse-mcp-python"
version = "1.0.0"
description = "Production-ready MCP server for Langfuse observability with comprehensive API coverage"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "langfuse>=2.0.0",
    "mcp>=0.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

agent-frameworks = [
    "langgraph>=0.2.0",
]

[project.scripts]
langfuse-monitor = "langfuse_mcp_python.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/langfuse_mcp_python"]
